% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assign_group_membership.R
\name{assign_group_membership}
\alias{assign_group_membership}
\title{Assign group membership to nodes}
\usage{
assign_group_membership(N, K, place_randomly = FALSE,
  ensure_all_groups_seen = TRUE)
}
\arguments{
\item{N}{Total number of nodes}

\item{K}{Total number of groups or clusters}

\item{place_randomly}{Do we want to randomly sample group membership for nodes? Defaults to \code{FALSE} so all groups are evenly (or as close to as possible) represented.}

\item{ensure_all_groups_seen}{If \code{place_randomly = TRUE}, do we make sure that at least one node is placed in every group?}
}
\value{
An array of memberships of length \code{N} made up of \code{K} groups as encoded with integers
}
\description{
Assign group membership to nodes
}
\examples{
assign_group_membership(10, 3)
assign_group_membership(10, 4, place_randomly = TRUE)
assign_group_membership(10, 4, place_randomly = TRUE, ensure_all_groups_seen = FALSE)
}
