% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_random_lambda.R
\name{generate_random_lambda}
\alias{generate_random_lambda}
\title{Generate a random lambda matrix}
\usage{
generate_random_lambda(K_a, K_b, random_generator = runif)
}
\arguments{
\item{K_a}{Number of clusters for node type a}

\item{K_b}{Number of clusters for node type b}

\item{random_generator}{Function used to generate average connection values. Defalts to just \code{runif}, needs to take a single argument of number of values to generate.}
}
\value{
Tibble with three columns: \code{a} for id of a node cluster, \code{b} for id of b node cluster, and \code{avg_num_cons} for the average number of connection between two nodes in cluster a and b respectively.
}
\description{
Given a \code{K_a} type a clusters, and \code{K_b} type b clusters and a random
generator for average number of connections between nodes, generates a tibble
that encodes all pairs of cluster types and the average number of connections
between them.
}
\examples{
generate_random_lambda(K_a = 3, K_b = 2, random_generator = function(n) rpois(n, 5))
}
