% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup_planted_pattern_model.R
\name{setup_planted_pattern_model}
\alias{setup_planted_pattern_model}
\title{Setup a planted pattern model}
\usage{
setup_planted_pattern_model(planted_patterns, num_noise_nodes = 10,
  noise_p = 0.2, planted_p_on = 0.95, planted_p_off = 0.05)
}
\arguments{
\item{planted_patterns}{Tibble containing planted patterns for node a groups
by activating b nodes with \code{1} or \code{0} in columns prefixed with pattern
\code{b1}, \code{b2}, .... Additionally a column \code{size} is used to control how many a
nodes have the planted pattern.}

\item{num_noise_nodes}{How many additional b nodes with no pattern are added
to model?}

\item{noise_p}{The individual lambda for connections between the noise
b-nodes and any a-node.}

\item{planted_p_on}{The lambda for a planted pattern b-node that has been
turned on in a pattern.}

\item{planted_p_off}{The lambda for a planted pattern b-node that has been
turned off in pattern.}
}
\value{
List with group membership vectors for both node types (\code{b_a, b_b}),
and the Lambda matrix for edge behavior (\code{Lambda}).
}
\description{
Supplied with a dataframe of planted patterns of b-nodes, will return all the
neccesary components to draw from SBM model with set planted structure using
\code{\link{draw_from_model()}}.
}
\examples{
my_patterns <- dplyr::tribble(
  ~b1,  ~b2,  ~b3,  ~b4,  ~size,
    1,    1,    0,    0,     10,
    1,    1,    1,    0,     15,
    0,    0,    0,    1,     13,
    0,    0,    1,    1,     24
)
setup_planted_pattern_model(my_patterns, num_noise_nodes = 15)

}
